Update service and controller to pass correct rest/html models
