--- src/main/java/net/saddlercoms/weatherviewer/WeatherService.java
@@ -1,5 +1,0 @@
-package net.saddlercoms.weatherviewer;
-
-public class WeatherService {
-
-}
